move Output enum to kafkactl.model & use COMPLETION-CANDIDATES variable for output help message & adjust test style
