+++ src/main/java/com/michelin/kafkactl/command/ResetPassword.java
@@ -1,8 +1,7 @@
 package com.michelin.kafkactl.command;
 
-import static com.michelin.kafkactl.service.FormatService.Output;
-
 import com.michelin.kafkactl.hook.AuthenticatedHook;
+import com.michelin.kafkactl.model.Output;
 import com.michelin.kafkactl.service.ResourceService;
 import io.micronaut.core.annotation.ReflectiveAccess;
 import jakarta.inject.Inject;
@@ -40,7 +39,7 @@ public class ResetPassword extends AuthenticatedHook {
 
     @Option(
         names = {"-o", "--output"},
-        description = "Output format. One of: yml|yaml|table",
+        description = "Output format (${COMPLETION-CANDIDATES}).",
         defaultValue = "table")
     public Output output;
