+++ src/main/java/com/michelin/kafkactl/model/Output.java
@@ -1,0 +1,22 @@
+package com.michelin.kafkactl.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+/**
+ * Output.
+ */
+@Getter
+@AllArgsConstructor
+public enum Output {
+    YAML("yaml"),
+    YML("yml"),
+    TABLE("table");
+
+    private final String name;
+
+    @Override
+    public String toString() {
+        return name;
+    }
+}
