+++ src/test/java/com/michelin/kafkactl/command/config/ConfigGetContextsTest.java
@@ -1,6 +1,6 @@
 package com.michelin.kafkactl.command.config;
 
-import static com.michelin.kafkactl.service.FormatService.Output.TABLE;
+import static com.michelin.kafkactl.model.Output.TABLE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.argThat;
