+++ src/test/java/com/michelin/kafkactl/service/FormatServiceTest.java
@@ -1,8 +1,8 @@
 package com.michelin.kafkactl.service;
 
-import static com.michelin.kafkactl.service.FormatService.Output.TABLE;
-import static com.michelin.kafkactl.service.FormatService.Output.YAML;
-import static com.michelin.kafkactl.service.FormatService.Output.YML;
+import static com.michelin.kafkactl.model.Output.TABLE;
+import static com.michelin.kafkactl.model.Output.YAML;
+import static com.michelin.kafkactl.model.Output.YML;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -171,7 +171,8 @@ class FormatServiceTest {
 
         formatService.displayList("RoleBinding", Collections.singletonList(resource), TABLE, cmd.getCommandSpec());
 
-        assertTrue(Pattern.compile("ROLE_BINDING\\s+GROUP\\s+VERBS\\s+RESOURCES").matcher(sw.toString()).find());
+        assertTrue(Pattern.compile("ROLE_BINDING\\s+GROUP\\s+VERBS\\s+RESOURCES")
+            .matcher(sw.toString()).find());
         assertTrue(Pattern.compile("roleBinding\\s+GROUP\\s+GET,POST,PUT,DELETE\\s+topics,acls,connectors")
             .matcher(sw.toString()).find());
     }
@@ -293,8 +294,10 @@ class FormatServiceTest {
 
         formatService.displaySingle(resource, TABLE, cmd.getCommandSpec());
 
-        assertTrue(Pattern.compile("TOPIC\\s+RETENTION\\s+POLICY\\s+AGE").matcher(sw.toString()).find());
-        assertTrue(Pattern.compile("prefix.topic\\s+1m\\s+delete").matcher(sw.toString()).find());
+        assertTrue(Pattern.compile("TOPIC\\s+RETENTION\\s+POLICY\\s+AGE")
+            .matcher(sw.toString()).find());
+        assertTrue(Pattern.compile("prefix.topic\\s+1m\\s+delete")
+            .matcher(sw.toString()).find());
 
     }
